+++ spyne/protocol/dictdoc.py
@@ -194,10 +194,6 @@ class FlatDictDocument(DictDocument):
                 else:
                     native_v2 = self.from_string(member.type, v2)
 
-                if (validator is self.SOFT_VALIDATION and native_v2 is None
-                                  and member.type.Attributes.nullable == False):
-                    raise ValidationError(native_v2)
-
                 if (validator is self.SOFT_VALIDATION and not
                             member.type.validate_native(member.type, native_v2)):
                     raise ValidationError(v2)
@@ -407,10 +403,6 @@ class HierDictDocument(DictDocument):
 
             else:
                 retval = self.from_string(class_, value)
-
-        if (validator is self.SOFT_VALIDATION and retval is None
-                          and class_.Attributes.nullable == False):
-            raise ValidationError(retval)
 
         # validate native type
         if validator is self.SOFT_VALIDATION and \